Event queue rework (#3)
Switch to esp's default queue instead of discreet queues so multiple consumers can more easily handle one message.

+++ main/Screen.h
@@ -5,6 +5,7 @@
 #include <u8g2.h>
 #include <memory>
 #include "shared.h"
+#include "state.h"
 
 extern "C"
 {
@@ -28,8 +29,8 @@ class Screen
 	void ToggleUnits();
 
   private:
-	uint16_t mySpeed;
-	uint16_t myPrevSpeed;
+	uint32_t mySpeed;
+	uint32_t myPrevSpeed;
 	UIState myState;
 	UIState myPrevState;
 	SpeedState mySpeedState;
